Split ContainerScanStepRunner functionality into several classes
